requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# This file was created with: conda list -e > requirements.txt
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.5=py311ha68e1ae_0
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
blosc=1.21.4=hdccc3a2_0
brotli=1.0.9=hcfcfb64_9
brotli-bin=1.0.9=hcfcfb64_9
brotli-python=1.0.9=py311h12c1d0e_9
bzip2=1.0.8=h8ffe710_4
c-ares=1.19.1=hcfcfb64_0
ca-certificates=2023.7.22=h56e8100_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h7d9ee11_3
cftime=1.6.2=py311h59ca53f_1
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=8.1.6=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.0=py311h005e61a_0
cryptography=41.0.3=py311h28e9c30_0
cuda-cccl=12.2.128=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
curl=8.1.1=h2bbff1b_0
cycler=0.11.0=pyhd8ed1ab_0
filelock=3.12.2=pyhd8ed1ab_0
fonttools=4.42.0=py311ha68e1ae_0
freetype=2.12.1=h546665d_1
frozenlist=1.4.0=py311ha68e1ae_0
fsspec=2023.6.0=pyh1a96a4e_0
gettext=0.21.1=h5728263_0
glib=2.76.4=h12be248_0
glib-tools=2.76.4=h12be248_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
grpcio=1.57.0=py311h5bc0dda_0
gst-plugins-base=1.22.5=h001b923_0
gstreamer=1.22.5=hb4038d2_0
hdf4=4.2.15=h1334946_6
hdf5=1.14.0=nompi_h918d9b7_103
icu=72.1=h63175ca_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
intel-openmp=2023.2.0=h57928b3_49496
jinja2=3.1.2=pyhd8ed1ab_1
kiwisolver=1.4.4=py311h005e61a_1
krb5=1.20.1=heb0366b_0
lame=3.100=hcfcfb64_1003
lcms2=2.15=h3e3b177_1
lerc=4.0.0=h63175ca_0
libabseil=20230125.3=cxx17_h63175ca_0
libaec=1.0.6=h63175ca_1
libbrotlicommon=1.0.9=hcfcfb64_9
libbrotlidec=1.0.9=hcfcfb64_9
libbrotlienc=1.0.9=hcfcfb64_9
libclang=16.0.6=default_heb8d277_1
libclang13=16.0.6=default_hc80b9e7_1
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.129=0
libcurand-dev=10.3.3.129=0
libcurl=8.1.1=h86230a5_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.18=hcfcfb64_0
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libflac=1.4.3=h63175ca_0
libglib=2.76.4=he8f3873_0
libgrpc=1.57.0=hea2d5f7_0
libhwloc=2.9.2=nocuda_h15da153_1008
libiconv=1.17=h8ffe710_0
libjpeg-turbo=2.1.5.1=hcfcfb64_0
libnetcdf=4.9.2=nompi_ha5afab8_105
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h19919ed_0
libprotobuf=4.23.3=h1975477_0
libsndfile=1.2.0=h2628c91_0
libsqlite=3.42.0=hcfcfb64_0
libssh2=1.11.0=h7dfc565_0
libtiff=4.5.1=h6c8260b_0
libuv=1.44.2=hcfcfb64_1
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.1=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.11.4=hc3477c8_0
libzip=1.9.2=h519de47_1
libzlib=1.2.13=hcfcfb64_5
lightning-utilities=0.9.0=pyhd8ed1ab_0
lz4-c=1.9.4=hcfcfb64_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.4=pyhd8ed1ab_0
markupsafe=2.1.3=py311ha68e1ae_0
matplotlib=3.7.2=py311h1ea47a8_0
matplotlib-base=3.7.2=py311h6e989c2_0
mkl=2023.2.0=h6a75c08_49573
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
mpg123=1.31.3=h63175ca_0
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
multidict=6.0.4=py311ha68e1ae_0
munkres=1.1.4=pyh9f0ad1d_0
netcdf4=1.6.4=nompi_py311h896b27b_100
networkx=3.1=pyhd8ed1ab_0
ninja=1.11.1=h91493d7_0
numpy=1.25.2=py311hdab7c0b_0
numpy-base=1.25.2=py311hd01c5d8_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.1.2=hcfcfb64_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py311hf63dbb6_1
pcre2=10.40=h17e33f8_0
pillow=10.0.0=py311hde623f7_0
pip=23.2.1=pyhd8ed1ab_0
platformdirs=3.10.0=pyhd8ed1ab_0
ply=3.11=py_1
protobuf=4.23.3=py311h03b55d4_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.9=py311h125bc19_4
pyqt5-sip=12.12.2=py311h12c1d0e_4
pysocks=1.7.1=pyh0701188_6
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.11.4=h2628c8c_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=3_cp311
pytorch=2.0.1=py3.11_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-lightning=2.0.6=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py311ha68e1ae_5
qt-main=5.15.8=h2c8576c_12
re2=2023.03.02=hd4eee63_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
setuptools=68.0.0=pyhd8ed1ab_0
sip=6.7.11=py311h12c1d0e_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=hfb803bf_0
sympy=1.12=pyh04b8f61_3
tbb=2021.10.0=h91493d7_0
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py311h1ea47a8_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.0.2=pypi_0
torchmetrics=1.0.3=pyhd8ed1ab_0
torchvision=0.15.2=cpu_py311haf6e6b9_0
tornado=6.3.2=py311ha68e1ae_0
tqdm=4.65.2=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.0.4=pyhd8ed1ab_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hfdfe4a8_17
vs2015_runtime=14.36.32532=h05e6639_17
werkzeug=2.3.6=pyhd8ed1ab_0
wheel=0.41.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xarray=2023.6.0=py311haa95532_0
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
yarl=1.9.2=py311ha68e1ae_0
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.2=h12be248_7
